before kiddy

+++ src/main/java/Domain/Entity/SystemAdmin.java
@@ -3,6 +3,8 @@ package Domain.Entity;
 import org.springframework.stereotype.Component;
 
 import javax.persistence.*;
+import java.util.ArrayList;
+import java.util.List;
 
 @Component
 @Entity(name="SystemAdmins")
@@ -11,7 +13,8 @@ public class SystemAdmin {
     @Id
     @Column(name = "Id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long systemAdminId;
+    private int systemAdminId;
+
 
 
 }
